Changed progress bar from referencing a hardcoded thread count to referencing a list of progress-influencing threads.
This is now done by a flag to set a thread as a progress thread, and tracking a list of them. Instead of checking if there's more than a total number of threads running, the progress class now checks to see if there's more than one (the thread triggering the progress bar) thread in that list that's still alive.

This lets us get away from hardcoding a thread counter, and allows us to easily have threads like the update check that don't count towards that progress bar counter.