Fixed window positioning in Linux
In Windows, winfo_y is the top left corner of the window. For some reason, in Linux, this doesn't account for the titlebar or menubar height. This means that in Linux, when setting Y in geometry() for some arbitrary Y position, if you get that back with winfo_y, or winfo_rooty, the return value is the geometry() value that was set, plus the height of the title and menubar.

The fix for this is to just get the geometry value, and parse the X and Y from there, avoiding winfo_ entirely.