Break analysis source distribution into chunks
This uses the same thing that was done for files in source folders, and splits the source list into chunks of at most 15 items. This prevents basically infinitely long wait times while the program brute forces all the permutations of files in the list to find the best way to split things.