Fix regression with source drive selection
When selecting the drive BackDrop is currently running from as the source drive in single drive selection mode, the folder list reads from the working directory that BackDrop runs out of, rather than the root of the drive. This was caused by os.walk() reading PWD instead of the drive root if you omit the trailing backslash (e.g "D:" instead of "D:\").